test: mudando estrutura de testes para melhorar o coverage
