+++ app/Http/Controllers/Api/Auth/AuthenticateController.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Auth;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\LoginRequest;
+use App\Http\Requests\Auth\LoginRequest;
 use App\Services\Authenticate\AuthenticateService;
 use Illuminate\Http\JsonResponse;
 
@@ -20,29 +20,11 @@ class AuthenticateController extends Controller
         $this->service = $service;
     }
 
-    /**
-     * @route POST /api/v1/login
-     * @tags Authentication
-     * @title Logar um usuário
-     * @description Pass user credentials
-     * @bodyParam email string required The user's email address
-     * @bodyParam password string required The user's password
-     * @response 200 Success
-     * @response 401 Unauthorized
-     */
     public function login(LoginRequest $request): JsonResponse
     {
         return response()->json($this->service->login($request->validated()));
     }
 
-    /**
-     * @route POST /api/v1/logout
-     * @tags Authentication
-     * @title Deslogar um usuário
-     * @authenticated
-     * @response 200 Success
-     * @response 200 { "message": "Logout feito com sucesso!" }
-     */
     public function logout(): JsonResponse
     {
         return response()->json([
@@ -50,9 +32,6 @@ class AuthenticateController extends Controller
         ]);
     }
 
-    /**
-     * @authenticated
-     */
     public function myProfile(): JsonResponse
     {
         return response()->json($this->service->myProfile());
