+++ app/Http/Controllers/Api/Permission/PermissionController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Permission;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PermissionResource;
