+++ tests/Feature/Http/Controller/Api/Permission/CheckPermissionTest.php
